refactor(cpp): give packet the socket that received it instead of socket->handle_packet
