+++ platforms/linux/src/Application/Packets/Commands/FindInformation/Peripheral/FindInformation.cpp
@@ -1,6 +1,6 @@
-#include <algorithm>
-#include "utils/string_formats.hpp"
 #include "FindInformation.hpp"
+#include "utils/string_formats.hpp"
+#include "Transport/Socket/HCI/HCISocket.hpp"
 
 using namespace std;
 
@@ -18,21 +18,59 @@ namespace Packet {
         m_error = Format::HCI::AttributeErrorCode::None;
       }
 
-      void FindInformation::set_gatt_table(const std::vector<Format::HCI::Service>& services, const std::vector<Format::HCI::Characteristic>& characteristics) {
+      vector<uint8_t> FindInformation::serialize(HCIFormatBuilder& builder) const {
+        if (m_error != Format::HCI::AttributeErrorCode::None) {
+          builder
+              .set_opcode(Format::HCI::AttributeCode::ErrorResponse)
+              .add(static_cast<uint8_t>(initial_packet_code()))
+              .add(m_starting_handle)
+              .add(m_error);
+        } else {
+          ControllerOnlyPacket::serialize(builder);
+
+          builder.add(m_format);
+
+          for (auto& info : m_info) {
+            builder
+                .add(info.first)
+                .add(info.second);
+          }
+        }
+
+        return builder.build(Format::HCI::Type::AsyncData);
+      }
+
+      void FindInformation::unserialize(HCIFormatExtractor& extractor) {
+        m_starting_handle = extractor.get_value<uint16_t>();
+        m_ending_handle = extractor.get_value<uint16_t>();
+
+        if (m_starting_handle < 0x0001 || m_starting_handle > m_ending_handle) {
+          LOG.warning("Invalid handle received: starting_handle=" + to_string(m_starting_handle) + ", ending_handle=" + to_string(m_ending_handle));
+          m_error = Format::HCI::AttributeErrorCode::InvalidHandle;
+          return;
+        }
+      }
+
+      void FindInformation::set_socket(AbstractSocket* socket) {
+        auto hci_socket = dynamic_cast<HCISocket*>(socket);
+        if (hci_socket == nullptr) {
+          throw Exceptions::BaBLEException(BaBLE::StatusCode::Failed, "Can't downcast socket to HCISocket packet");
+        }
+
         if (m_error != Format::HCI::AttributeErrorCode::None) {
           return;
         }
 
-        for (auto& service : services) {
+        for (auto& service : hci_socket->get_services()) {
           if (service.handle >= m_starting_handle && service.handle <= m_ending_handle) {
             m_info[service.handle] = vector<uint8_t>{
-              static_cast<uint8_t>(Format::HCI::GattUUID::PrimaryServiceDeclaration & 0x00FF),
-              static_cast<uint8_t>(Format::HCI::GattUUID::PrimaryServiceDeclaration >> 8)
+                static_cast<uint8_t>(Format::HCI::GattUUID::PrimaryServiceDeclaration & 0x00FF),
+                static_cast<uint8_t>(Format::HCI::GattUUID::PrimaryServiceDeclaration >> 8)
             };
           }
         }
 
-        for (auto& characteristic : characteristics) {
+        for (auto& characteristic : hci_socket->get_characteristics()) {
           if (characteristic.handle >= m_starting_handle && characteristic.handle <= m_ending_handle) {
             m_info[characteristic.handle] = vector<uint8_t>{
                 static_cast<uint8_t>(Format::HCI::GattUUID::CharacteristicDeclaration & 0x00FF),
@@ -79,39 +117,6 @@ namespace Packet {
         }
 
         m_info.erase(it, m_info.end());
-      }
-
-      vector<uint8_t> FindInformation::serialize(HCIFormatBuilder& builder) const {
-        if (m_error != Format::HCI::AttributeErrorCode::None) {
-          builder
-              .set_opcode(Format::HCI::AttributeCode::ErrorResponse)
-              .add(static_cast<uint8_t>(initial_packet_code()))
-              .add(m_starting_handle)
-              .add(m_error);
-        } else {
-          ControllerOnlyPacket::serialize(builder);
-
-          builder.add(m_format);
-
-          for (auto& info : m_info) {
-            builder
-                .add(info.first)
-                .add(info.second);
-          }
-        }
-
-        return builder.build(Format::HCI::Type::AsyncData);
-      }
-
-      void FindInformation::unserialize(HCIFormatExtractor& extractor) {
-        m_starting_handle = extractor.get_value<uint16_t>();
-        m_ending_handle = extractor.get_value<uint16_t>();
-
-        if (m_starting_handle < 0x0001 || m_starting_handle > m_ending_handle) {
-          LOG.warning("Invalid handle received: starting_handle=" + to_string(m_starting_handle) + ", ending_handle=" + to_string(m_ending_handle));
-          m_error = Format::HCI::AttributeErrorCode::InvalidHandle;
-          return;
-        }
       }
 
       const string FindInformation::stringify() const {
