+++ platforms/linux/src/Application/Packets/Commands/Read/Peripheral/ReadResponse.cpp
@@ -1,0 +1,9 @@
+#include "ReadResponse.hpp"
+
+ReadResponse::ReadResponse() {
+
+}
+
+ReadResponse::~ReadResponse() {
+
+}
