+++ platforms/linux/src/Application/Packets/Commands/Read/Peripheral/ReadResponse.hpp
@@ -1,0 +1,14 @@
+#ifndef BABLE_READRESPONSE_HPP
+#define BABLE_READRESPONSE_HPP
+
+class ReadResponse {
+
+public:
+  ReadResponse();
+  ~ReadResponse();
+
+private:
+
+};
+
+#endif //BABLE_READRESPONSE_HPP
