+++ platforms/linux/src/Application/Packets/Commands/ReadByGroupType/Peripheral/ReadByGroupType.hpp
@@ -29,9 +29,9 @@ namespace Packet {
         void unserialize(HCIFormatExtractor& extractor) override;
         std::vector<uint8_t> serialize(HCIFormatBuilder& builder) const override;
 
-        const std::string stringify() const override;
+        void set_socket(AbstractSocket* socket) override;
 
-        void set_services(const std::vector<Format::HCI::Service>& services);
+        const std::string stringify() const override;
 
       private:
         uint16_t m_starting_handle;
