+++ platforms/linux/src/Application/Packets/Events/DeviceDisconnected/DeviceDisconnected.hpp
@@ -27,6 +27,8 @@ namespace Packet {
       void unserialize(HCIFormatExtractor& extractor) override;
       std::vector<uint8_t> serialize(FlatbuffersFormatBuilder& builder) const override;
 
+      void set_socket(AbstractSocket* socket) override;
+
       const std::string stringify() const override;
 
     private:
