+++ platforms/linux/tests/test_integration.cpp
@@ -6,8 +6,8 @@
 #include "Application/Packets/Meta/GetControllersList/GetControllersList.hpp"
 #include "Application/Packets/Commands/GetControllersIds/GetControllersIdsResponse.hpp"
 #include "Application/Packets/Commands/GetControllerInfo/GetControllerInfoResponse.hpp"
-#include "Application/Packets/Commands/Read/ReadRequest.hpp"
-#include "Application/Packets/Commands/Read/ReadResponse.hpp"
+#include "Application/Packets/Commands/Read/Central/ReadRequest.hpp"
+#include "Application/Packets/Commands/Read/Central/ReadResponse.hpp"
 #include "Application/Packets/Commands/Write/WriteResponse.hpp"
 #include "Application/Packets/Control/Ready/Ready.hpp"
 #include "mocks/mock_socket.hpp"
@@ -30,6 +30,10 @@ TEST_CASE("Integration (with mocked socket) - MGMT meta packet", "[integration][
   // Mocked sockets
   shared_ptr<MockMGMTSocket> mgmt_socket = make_shared<MockMGMTSocket>(loop, mgmt_format);
   shared_ptr<MockStdIOSocket> stdio_socket = make_shared<MockStdIOSocket>(loop, fb_format);
+  REQUIRE(mgmt_socket->get_raw()->is_open());
+  REQUIRE(mgmt_socket->get_raw()->is_option_set() == false);
+  REQUIRE(mgmt_socket->get_raw()->is_binded());
+  REQUIRE(mgmt_socket->get_raw()->is_connected() == false);
 
   // Socket container
   SocketContainer socket_container;
@@ -170,6 +174,10 @@ TEST_CASE("Integration (with mocked socket) - HCI packet", "[integration][hci]")
   // Mocked sockets
   shared_ptr<MockStdIOSocket> stdio_socket = make_shared<MockStdIOSocket>(loop, fb_format);
   shared_ptr<MockHCISocket> hci_socket = make_shared<MockHCISocket>(loop, hci_format, 0, Utils::extract_bd_address(controller_address));
+  REQUIRE(hci_socket->get_raw()->is_open());
+  REQUIRE(hci_socket->get_raw()->is_option_set());
+  REQUIRE(hci_socket->get_raw()->is_binded());
+  REQUIRE(hci_socket->get_raw()->is_connected() == false);
   REQUIRE(hci_socket->get_controller_id() == 0);
   REQUIRE(hci_socket->get_controller_address() == controller_address);
 
@@ -185,12 +193,12 @@ TEST_CASE("Integration (with mocked socket) - HCI packet", "[integration][hci]")
   // PacketBuilder
   PacketBuilder hci_packet_builder(hci_format);
   hci_packet_builder
-      .register_command<Packet::Commands::ReadResponse>()
+      .register_command<Packet::Commands::Central::ReadResponse>()
       .register_command<Packet::Commands::WriteResponse>();
 
   PacketBuilder stdio_packet_builder(fb_format);
   stdio_packet_builder
-      .register_command<Packet::Commands::ReadRequest>();
+      .register_command<Packet::Commands::Central::ReadRequest>();
 
   auto on_error = [&stdio_socket, &socket_container](const Exceptions::BaBLEException& err) {
     CAPTURE(err);
@@ -228,7 +236,7 @@ TEST_CASE("Integration (with mocked socket) - HCI packet", "[integration][hci]")
   );
 
   // Test send ReadRequest (HCI packet) and receive result
-  shared_ptr<Packet::Commands::ReadRequest> read_request_packet = make_shared<Packet::Commands::ReadRequest>();
+  shared_ptr<Packet::Commands::Central::ReadRequest> read_request_packet = make_shared<Packet::Commands::Central::ReadRequest>();
   REQUIRE_THROWS(read_request_packet->to_bytes());
 
   // Build a ReadRequest
